pyproject.toml
[tool.poetry]
name = "updatenight-agent"
version = "0.1.0"
description = "Newsletter automation using crewAI"
authors = ["Your Name"]

[tool.poetry.dependencies]
python = "^3.11"
crewai = "^0.102.0"
crawl4ai = "^0.5.0.post4"
python-dotenv = "^1.0.1"
langchain = "^0.3.20"
langchain_community = "^0.3.19"
markdown = "^3.5.2"
discord = "^2.3.2"
pillow = "^10.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
